added FeedbackDTO

+++ src/main/java/com/eventManagement/EMS/service/FeedbackService.java
@@ -1,5 +1,6 @@
 package com.eventManagement.EMS.service;
 
+import com.eventManagement.EMS.DTO.FeedbackDTO;
 import com.eventManagement.EMS.models.Event;
 import com.eventManagement.EMS.models.Feedback;
 import com.eventManagement.EMS.models.User;
@@ -12,6 +13,7 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -37,15 +39,27 @@ public class FeedbackService {
     }
 
     //Displays all feedback of a specific event
-    public ResponseEntity<List<Feedback>> getAllEventFeedback(Long eventId){
+    public ResponseEntity<List<FeedbackDTO>> getAllEventFeedback(Long eventId){
         Optional<Event> eventOptional = eventRepository.findById(eventId);
 
         if (eventOptional.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-
         List<Feedback> feedbacks = feedbackRepository.findByEventId(eventId);
-        return new ResponseEntity<>(feedbacks, HttpStatus.OK);
+        List<FeedbackDTO> feedbackDTOList = new ArrayList<>();
+
+        for(Feedback feedback : feedbacks){
+            FeedbackDTO feedbackDTO = new FeedbackDTO();
+            feedbackDTO.setId(feedback.getId());
+            feedbackDTO.setComments(feedback.getComments());
+            feedbackDTO.setRate(feedback.getRate());
+            feedbackDTO.setEvent(feedback.getEvent().getName());
+            String fullName = feedback.getUser().getFirstName() + feedback.getUser().getLastName();
+            feedbackDTO.setUser(fullName);
+            feedbackDTO.setSubmittedAt(feedback.getSubmittedAt());
+            feedbackDTOList.add(feedbackDTO);
+        }
+        return new ResponseEntity<>(feedbackDTOList, HttpStatus.OK);
     }
 
     public ResponseEntity<String> editFeedback(Long feedbackId, Feedback updatedFeedback, User user){
@@ -57,7 +71,7 @@ public class FeedbackService {
         // Checks if the logged-in user is the one who wrote the feedback
         if(feedback.getUser().getUserID().equals(user.getUserID())){
             feedback.setComments(updatedFeedback.getComments() != null ? updatedFeedback.getComments() : feedback.getComments());
-            if(updatedFeedback.getRate() >= 0 && updatedFeedback != null){
+            if(updatedFeedback.getRate() >= 0){
                 feedback.setRate(updatedFeedback.getRate());
             }else{
                 feedback.setRate(feedback.getRate());
@@ -71,9 +85,16 @@ public class FeedbackService {
 
     public ResponseEntity<String> deleteFeedback(Long feedbackId, User user){
         Optional<Feedback> feedbackOptional = feedbackRepository.findById(feedbackId);
-        Feedback feedback = feedbackOptional.get();
-        feedbackRepository.delete(feedback);
-        return new ResponseEntity<>("Feedback has been deleted", HttpStatus.NO_CONTENT);
-    }
 
+        if(feedbackOptional.isEmpty()){
+            return new ResponseEntity<>("Feedback not found", HttpStatus.NOT_FOUND);
+        }
+        Feedback feedback = feedbackOptional.get();
+
+        if(user.getUserID().equals(feedback.getUser().getUserID())){
+            feedbackRepository.delete(feedback);
+            return new ResponseEntity<>("Feedback has been deleted", HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>("You can't delete this comment", HttpStatus.UNAUTHORIZED);
+    }
 }
